Renames sandbox-cache to virtual-node-cache, as agreed.
