+++ libs/flows/flow-manager/build.gradle
@@ -16,5 +16,5 @@ dependencies {
     implementation "org.apache.servicemix.bundles:org.apache.servicemix.bundles.kafka-clients:$kafkaClientVersion"
 
     implementation project(':libs:messaging:messaging')
-    implementation project(':libs:sandbox:sandbox-cache')
+    implementation project(':libs:virtual-node:virtual-node-cache')
 }
