--- libs/sandbox/sandbox-cache-impl/build.gradle
@@ -1,24 +1,0 @@
-plugins {
-    id 'com.r3.internal.gradle.plugins.r3Publish'
-    id 'corda.common-library'
-}
-
-description "Sandbox Cache Implementation"
-
-dependencies {
-    compileOnly "org.osgi:osgi.annotation:$osgiVersion"
-    compileOnly "org.osgi:org.osgi.service.component.annotations:$osgiScrAnnotationVersion"
-
-    implementation platform("net.corda:corda-api:$cordaApiVersion")
-    implementation "net.corda:corda-application"
-    implementation "net.corda:corda-base"
-    implementation "net.corda:corda-install:$cordaVersion"
-    implementation "net.corda:corda-packaging:$cordaVersion"
-    implementation "net.corda:corda-sandbox:$cordaVersion"
-    implementation "net.corda:avro-schema"
-
-    implementation project(":libs:sandbox:sandbox-cache")
-
-    testImplementation "org.assertj:assertj-core:$assertjVersion"
-    testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
-}
