--- libs/sandbox/sandbox-cache-impl/src/main/java/net/corda/sandbox/cache/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.sandbox.cache;
-
-import org.osgi.annotation.bundle.Export;
