+++ libs/virtual-node/virtual-node-cache-impl/src/main/java/net/corda/virtual/node/cache/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.virtual.node.cache;
+
+import org.osgi.annotation.bundle.Export;
