+++ settings.gradle
@@ -98,8 +98,6 @@ include 'libs:messaging:db-messaging-impl'
 include 'libs:messaging:kafka-messaging-impl'
 include 'libs:messaging:inmemory-messaging-impl'
 include 'libs:p2p-crypto'
-include 'libs:sandbox:sandbox-cache'
-include 'libs:sandbox:sandbox-cache-impl'
 include 'libs:schema-registry:schema-registry'
 include 'libs:schema-registry:schema-registry-impl'
 include 'libs:security-manager'
@@ -114,6 +112,8 @@ include 'libs:serialization:serialization-amqp'
 include 'libs:serialization:serialization-encoding'
 include 'libs:serialization:serialization-kryo'
 include 'libs:utilities'
+include 'libs:virtual-node:virtual-node-cache'
+include 'libs:virtual-node:virtual-node-cache-impl'
 include 'libs:http-rpc:http-rpc'
 include 'osgi-framework-api'
 include 'osgi-framework-bootstrap'
